+++ index.js
@@ -722,28 +722,42 @@ class Ofn {
     //endregion
     //region Dates
 
-    static dateObjByToday( sep = '/' ) {
-        return Ofn.dateObjByDate( (new Date()), sep );
+    static dateObjByToday( opts = {} ) {
+        Ofn.isString( opts ) && ( opts = { sep: opts } );
+        ! Ofn.isObject( opts ) && ( opts = { sep: '/' } );
+        ! Ofn.isString( opts.sep ) && ( opts.sep = '/' );
+        return Ofn.dateObjByDate( (new Date()), opts );
     }
 
-    static dateObjByTimestamp( timestamp, sep = '/' ) {
+    static dateObjByTimestamp( timestamp, opts = {} ) {
         if( ! Ofn.isNumber( timestamp ) ) { return null; }
 
-        return Ofn.dateObjByDate( (new Date( timestamp )), sep );
+        Ofn.isString( opts ) && ( opts = { sep: opts } );
+        ! Ofn.isObject( opts ) && ( opts = { sep: '/' } );
+        ! Ofn.isString( opts.sep ) && ( opts.sep = '/' );
+        return Ofn.dateObjByDate( (new Date( timestamp )), opts );
     }
 
-    static dateObjBySql( sqldate, sep = '/' ) {
+    static dateObjBySql( sqldate, opts = {} ) {
         if( ! Ofn.isString( sqldate ) ) { return null; }
         if( ! REGEXP.DATETIME_SQL.test( sqldate ) ) { return null; }
 
-        return Ofn.dateObjByDate( (new Date( sqldate )), sep );
+        Ofn.isString( opts ) && ( opts = { sep: opts } );
+        ! Ofn.isObject( opts ) && ( opts = { sep: '/' } );
+        ! Ofn.isString( opts.sep ) && ( opts.sep = '/' );
+        return Ofn.dateObjByDate( (new Date( sqldate )), opts );
     }
 
-    static dateObjByHtml( html, sep = '/' ) {
+    static dateObjByHtml( html, opts = {} ) {
         if( ! Ofn.isString( html ) ) { return null; }
-        if( ! REGEXP.DATETIME_FN( sep ).test( html ) ) { return null; }
 
-        const arr = html.trim().split( sep );
+        Ofn.isString( opts ) && ( opts = { sep: opts } );
+        ! Ofn.isObject( opts ) && ( opts = { sep: '/' } );
+        ! Ofn.isString( opts.sep ) && ( opts.sep = '/' );
+
+        if( ! REGEXP.DATETIME_FN( opts.sep ).test( html ) ) { return null; }
+
+        const arr = html.trim().split( opts.sep );
         if( arr.length !== 3 ) { return null; }
 
         const yearhours = arr[ 2 ].split( ' ' );
@@ -752,31 +766,38 @@ class Ofn {
             sqldate += ` ${yearhours[ 1 ]}`;
         }
 
-        return Ofn.dateObjByDate( (new Date( sqldate )), sep );
+        return Ofn.dateObjByDate( (new Date( sqldate )), opts );
     }
 
-    static dateObjByDate( date, sep = '/' ) {
-        if( Ofn.type( date ) !== 'date' ) {
-            return null;
-        }
+    static dateObjByDate( date, opts = {} ) {
+        if( Ofn.type( date ) !== 'date' ) { return null; }
 
         let dateObj = {};
         dateObj.iso = date.toISOString();
 
-        let localData = new Date( date );
-        localData.setHours( localData.getHours() + ( localData.getTimezoneOffset() / -60 ) );
-        dateObj.local = localData.toISOString();
-        dateObj.sep = sep;
+        //
 
-        const dateArr = dateObj.local.split( /[^0-9]/ ).slice( 0, -1 );
-        // dateArr = [ year, month, day, hour, minute, second, millisecond ]
+        Ofn.isString( opts ) && ( opts = { sep: opts } );
+        ! Ofn.isObject( opts ) && ( opts = { sep: '/' } );
+        ! Ofn.isString( opts.sep ) && ( opts.sep = '/' );
 
-        dateObj.year = dateArr[ 0 ];
+        if( Ofn.isString( opts.timezone ) && ! isTimezoneAllowed( opts.timezone ) ) { return null; }
+
+        dateObj.timezone = opts.timezone ? opts.timezone : Intl.DateTimeFormat().resolvedOptions().timeZone
+        dateObj.sep = opts.sep;
+
+        let millis = Ofn.strPad( date.getMilliseconds(), 3, '0', 'right' );
+        const dateArr = date.toLocaleString( 'en-GB', { timeZone: dateObj.timezone } ).split( /[^0-9]/ );
+        // dateArr = [ day, month, year, empty, hour, minute, second ]
+
+        dateObj.local = `${dateArr[ 2 ]}-${dateArr[ 1 ]}-${dateArr[ 0 ]}T${dateArr[ 4 ]}:${dateArr[ 5 ]}:${dateArr[ 6 ]}.${millis}Z`;
+
+        dateObj.year = dateArr[ 2 ];
         dateObj.month = dateArr[ 1 ];
-        dateObj.day = dateArr[ 2 ];
-        dateObj.hour = dateArr[ 3 ];
-        dateObj.minute = dateArr[ 4 ];
-        dateObj.second = dateArr[ 5 ];
+        dateObj.day = dateArr[ 0 ];
+        dateObj.hour = dateArr[ 4 ];
+        dateObj.minute = dateArr[ 5 ];
+        dateObj.second = dateArr[ 6 ];
 
         dateObj.time = `${dateObj.hour}:${dateObj.minute}`;
         dateObj.times = `${dateObj.hour}:${dateObj.minute}:${dateObj.second}`;
@@ -787,13 +808,14 @@ class Ofn {
 
         dateObj.sqldate = `${dateObj.year}-${dateObj.month}-${dateObj.day}`;
         dateObj.sqldatetime = `${dateObj.sqldate} ${dateObj.times}`;
+        dateObj.sqldatetimeIso = dateObj.iso.replace( 'T', ' ' ).substring( 0, 19 );
 
         dateObj.fulldate = `${dateObj.date} ${dateObj.times}`;
         dateObj.fulldateStart = `${dateObj.date} 00:00:00`;
         dateObj.fulldateEnd = `${dateObj.date} 23:59:59`;
 
         dateObj.weekday = [ 7, 1, 2, 3, 4, 5, 6 ][ date.getDay() ];
-        dateObj.monthdays = new Date( dateObj.year, dateObj.month, 0 ).getDate();
+        dateObj.monthdays = new Date( +dateObj.year, +dateObj.month, 0 ).getDate();
         dateObj.timestamp = date.getTime();
 
         dateObj.jsdate = new Date( date );
@@ -802,14 +824,13 @@ class Ofn {
 
     static dateObjPlusDays( date, days ) {
         if( ! Ofn.isDate( date ) && ! Ofn.dateIsObj( date ) ) { return null; }
-        if( ! Ofn.isNumeric( days ) && Ofn.dateIsObj( date ) ) { return Ofn.cloneObject( date ); }
 
         days = ! Ofn.isNumeric( days ) ? 0 : +days;
 
         let jsdate = new Date( Ofn.dateIsObj( date ) ? date.jsdate : date );
         jsdate.setDate( jsdate.getDate() + days );
 
-        return Ofn.dateObjByDate( jsdate, Ofn.dateIsObj( date ) ? date.sep : '/' )
+        return Ofn.dateObjByDate( jsdate, Ofn.dateIsObj( date ) ? date : { sep: '/' } );
     }
 
     static dateIsObj( date ) {
@@ -1294,4 +1315,364 @@ class Ofn {
     //endregion
 }
 
+function isTimezoneAllowed( timezone ) {
+    if( ! Ofn.isString( timezone ) ) { return false; }
+
+    const TIMEZONES = [
+        'Africa/Abidjan',
+        'Africa/Accra',
+        'Africa/Algiers',
+        'Africa/Bissau',
+        'Africa/Cairo',
+        'Africa/Casablanca',
+        'Africa/Ceuta',
+        'Africa/El_Aaiun',
+        'Africa/Johannesburg',
+        'Africa/Juba',
+        'Africa/Khartoum',
+        'Africa/Lagos',
+        'Africa/Maputo',
+        'Africa/Monrovia',
+        'Africa/Nairobi',
+        'Africa/Ndjamena',
+        'Africa/Sao_Tome',
+        'Africa/Tripoli',
+        'Africa/Tunis',
+        'Africa/Windhoek',
+        'America/Adak',
+        'America/Anchorage',
+        'America/Araguaina',
+        'America/Argentina/Buenos_Aires',
+        'America/Argentina/Catamarca',
+        'America/Argentina/Cordoba',
+        'America/Argentina/Jujuy',
+        'America/Argentina/La_Rioja',
+        'America/Argentina/Mendoza',
+        'America/Argentina/Rio_Gallegos',
+        'America/Argentina/Salta',
+        'America/Argentina/San_Juan',
+        'America/Argentina/San_Luis',
+        'America/Argentina/Tucuman',
+        'America/Argentina/Ushuaia',
+        'America/Asuncion',
+        'America/Atikokan',
+        'America/Bahia',
+        'America/Bahia_Banderas',
+        'America/Barbados',
+        'America/Belem',
+        'America/Belize',
+        'America/Blanc-Sablon',
+        'America/Boa_Vista',
+        'America/Bogota',
+        'America/Boise',
+        'America/Cambridge_Bay',
+        'America/Campo_Grande',
+        'America/Cancun',
+        'America/Caracas',
+        'America/Cayenne',
+        'America/Chicago',
+        'America/Chihuahua',
+        'America/Costa_Rica',
+        'America/Creston',
+        'America/Cuiaba',
+        'America/Curacao',
+        'America/Danmarkshavn',
+        'America/Dawson',
+        'America/Dawson_Creek',
+        'America/Denver',
+        'America/Detroit',
+        'America/Edmonton',
+        'America/Eirunepe',
+        'America/El_Salvador',
+        'America/Fort_Nelson',
+        'America/Fortaleza',
+        'America/Glace_Bay',
+        'America/Godthab',
+        'America/Goose_Bay',
+        'America/Grand_Turk',
+        'America/Guatemala',
+        'America/Guayaquil',
+        'America/Guyana',
+        'America/Halifax',
+        'America/Havana',
+        'America/Hermosillo',
+        'America/Indiana/Indianapolis',
+        'America/Indiana/Knox',
+        'America/Indiana/Marengo',
+        'America/Indiana/Petersburg',
+        'America/Indiana/Tell_City',
+        'America/Indiana/Vevay',
+        'America/Indiana/Vincennes',
+        'America/Indiana/Winamac',
+        'America/Inuvik',
+        'America/Iqaluit',
+        'America/Jamaica',
+        'America/Juneau',
+        'America/Kentucky/Louisville',
+        'America/Kentucky/Monticello',
+        'America/La_Paz',
+        'America/Lima',
+        'America/Los_Angeles',
+        'America/Maceio',
+        'America/Managua',
+        'America/Manaus',
+        'America/Martinique',
+        'America/Matamoros',
+        'America/Mazatlan',
+        'America/Menominee',
+        'America/Merida',
+        'America/Metlakatla',
+        'America/Mexico_City',
+        'America/Miquelon',
+        'America/Moncton',
+        'America/Monterrey',
+        'America/Montevideo',
+        'America/Nassau',
+        'America/New_York',
+        'America/Nipigon',
+        'America/Nome',
+        'America/Noronha',
+        'America/North_Dakota/Beulah',
+        'America/North_Dakota/Center',
+        'America/North_Dakota/New_Salem',
+        'America/Ojinaga',
+        'America/Panama',
+        'America/Pangnirtung',
+        'America/Paramaribo',
+        'America/Phoenix',
+        'America/Port-au-Prince',
+        'America/Port_of_Spain',
+        'America/Porto_Velho',
+        'America/Puerto_Rico',
+        'America/Punta_Arenas',
+        'America/Rainy_River',
+        'America/Rankin_Inlet',
+        'America/Recife',
+        'America/Regina',
+        'America/Resolute',
+        'America/Rio_Branco',
+        'America/Santarem',
+        'America/Santiago',
+        'America/Santo_Domingo',
+        'America/Sao_Paulo',
+        'America/Scoresbysund',
+        'America/Sitka',
+        'America/St_Johns',
+        'America/Swift_Current',
+        'America/Tegucigalpa',
+        'America/Thule',
+        'America/Thunder_Bay',
+        'America/Tijuana',
+        'America/Toronto',
+        'America/Vancouver',
+        'America/Whitehorse',
+        'America/Winnipeg',
+        'America/Yakutat',
+        'America/Yellowknife',
+        'Antarctica/Casey',
+        'Antarctica/Davis',
+        'Antarctica/DumontDUrville',
+        'Antarctica/Macquarie',
+        'Antarctica/Mawson',
+        'Antarctica/Palmer',
+        'Antarctica/Rothera',
+        'Antarctica/Syowa',
+        'Antarctica/Troll',
+        'Antarctica/Vostok',
+        'Asia/Almaty',
+        'Asia/Amman',
+        'Asia/Anadyr',
+        'Asia/Aqtau',
+        'Asia/Aqtobe',
+        'Asia/Ashgabat',
+        'Asia/Atyrau',
+        'Asia/Baghdad',
+        'Asia/Baku',
+        'Asia/Bangkok',
+        'Asia/Barnaul',
+        'Asia/Beirut',
+        'Asia/Bishkek',
+        'Asia/Brunei',
+        'Asia/Chita',
+        'Asia/Choibalsan',
+        'Asia/Colombo',
+        'Asia/Damascus',
+        'Asia/Dhaka',
+        'Asia/Dili',
+        'Asia/Dubai',
+        'Asia/Dushanbe',
+        'Asia/Famagusta',
+        'Asia/Gaza',
+        'Asia/Hebron',
+        'Asia/Ho_Chi_Minh',
+        'Asia/Hong_Kong',
+        'Asia/Hovd',
+        'Asia/Irkutsk',
+        'Asia/Jakarta',
+        'Asia/Jayapura',
+        'Asia/Jerusalem',
+        'Asia/Kabul',
+        'Asia/Kamchatka',
+        'Asia/Karachi',
+        'Asia/Kathmandu',
+        'Asia/Khandyga',
+        'Asia/Kolkata',
+        'Asia/Krasnoyarsk',
+        'Asia/Kuala_Lumpur',
+        'Asia/Kuching',
+        'Asia/Macau',
+        'Asia/Magadan',
+        'Asia/Makassar',
+        'Asia/Manila',
+        'Asia/Nicosia',
+        'Asia/Novokuznetsk',
+        'Asia/Novosibirsk',
+        'Asia/Omsk',
+        'Asia/Oral',
+        'Asia/Pontianak',
+        'Asia/Pyongyang',
+        'Asia/Qatar',
+        'Asia/Qostanay',
+        'Asia/Qyzylorda',
+        'Asia/Riyadh',
+        'Asia/Sakhalin',
+        'Asia/Samarkand',
+        'Asia/Seoul',
+        'Asia/Shanghai',
+        'Asia/Singapore',
+        'Asia/Srednekolymsk',
+        'Asia/Taipei',
+        'Asia/Tashkent',
+        'Asia/Tbilisi',
+        'Asia/Tehran',
+        'Asia/Thimphu',
+        'Asia/Tokyo',
+        'Asia/Tomsk',
+        'Asia/Ulaanbaatar',
+        'Asia/Urumqi',
+        'Asia/Ust-Nera',
+        'Asia/Vladivostok',
+        'Asia/Yakutsk',
+        'Asia/Yangon',
+        'Asia/Yekaterinburg',
+        'Asia/Yerevan',
+        'Atlantic/Azores',
+        'Atlantic/Bermuda',
+        'Atlantic/Canary',
+        'Atlantic/Cape_Verde',
+        'Atlantic/Faroe',
+        'Atlantic/Madeira',
+        'Atlantic/Reykjavik',
+        'Atlantic/South_Georgia',
+        'Atlantic/Stanley',
+        'Australia/Adelaide',
+        'Australia/Brisbane',
+        'Australia/Broken_Hill',
+        'Australia/Currie',
+        'Australia/Darwin',
+        'Australia/Eucla',
+        'Australia/Hobart',
+        'Australia/Lindeman',
+        'Australia/Lord_Howe',
+        'Australia/Melbourne',
+        'Australia/Perth',
+        'Australia/Sydney',
+        'Europe/Amsterdam',
+        'Europe/Andorra',
+        'Europe/Astrakhan',
+        'Europe/Athens',
+        'Europe/Belgrade',
+        'Europe/Berlin',
+        'Europe/Brussels',
+        'Europe/Bucharest',
+        'Europe/Budapest',
+        'Europe/Chisinau',
+        'Europe/Copenhagen',
+        'Europe/Dublin',
+        'Europe/Gibraltar',
+        'Europe/Helsinki',
+        'Europe/Istanbul',
+        'Europe/Kaliningrad',
+        'Europe/Kiev',
+        'Europe/Kirov',
+        'Europe/Lisbon',
+        'Europe/London',
+        'Europe/Luxembourg',
+        'Europe/Madrid',
+        'Europe/Malta',
+        'Europe/Minsk',
+        'Europe/Monaco',
+        'Europe/Moscow',
+        'Europe/Oslo',
+        'Europe/Paris',
+        'Europe/Prague',
+        'Europe/Riga',
+        'Europe/Rome',
+        'Europe/Samara',
+        'Europe/Saratov',
+        'Europe/Simferopol',
+        'Europe/Sofia',
+        'Europe/Stockholm',
+        'Europe/Tallinn',
+        'Europe/Tirane',
+        'Europe/Ulyanovsk',
+        'Europe/Uzhgorod',
+        'Europe/Vienna',
+        'Europe/Vilnius',
+        'Europe/Volgograd',
+        'Europe/Warsaw',
+        'Europe/Zaporozhye',
+        'Europe/Zurich',
+        'Indian/Chagos',
+        'Indian/Christmas',
+        'Indian/Cocos',
+        'Indian/Kerguelen',
+        'Indian/Mahe',
+        'Indian/Maldives',
+        'Indian/Mauritius',
+        'Indian/Reunion',
+        'Pacific/Apia',
+        'Pacific/Auckland',
+        'Pacific/Bougainville',
+        'Pacific/Chatham',
+        'Pacific/Chuuk',
+        'Pacific/Easter',
+        'Pacific/Efate',
+        'Pacific/Enderbury',
+        'Pacific/Fakaofo',
+        'Pacific/Fiji',
+        'Pacific/Funafuti',
+        'Pacific/Galapagos',
+        'Pacific/Gambier',
+        'Pacific/Guadalcanal',
+        'Pacific/Guam',
+        'Pacific/Honolulu',
+        'Pacific/Kiritimati',
+        'Pacific/Kosrae',
+        'Pacific/Kwajalein',
+        'Pacific/Majuro',
+        'Pacific/Marquesas',
+        'Pacific/Nauru',
+        'Pacific/Niue',
+        'Pacific/Norfolk',
+        'Pacific/Noumea',
+        'Pacific/Pago_Pago',
+        'Pacific/Palau',
+        'Pacific/Pitcairn',
+        'Pacific/Pohnpei',
+        'Pacific/Port_Moresby',
+        'Pacific/Rarotonga',
+        'Pacific/Tahiti',
+        'Pacific/Tarawa',
+        'Pacific/Tongatapu',
+        'Pacific/Wake',
+        'Pacific/Wallis',
+        'UTC'
+    ];
+
+    return TIMEZONES.includes( timezone );
+}
+
 module.exports = Ofn;
+
+